myApp/requirements.txt
click==8.1.3
colorama==0.4.6
cssmin==0.2.0
dnspython==2.3.0
Flask==2.2.3
Flask-HTMLmin==2.2.1
gunicorn==20.1.0
htmlmin==0.1.12
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
python-dotenv==1.0.0
webassets==2.0
Werkzeug==2.3.3
